test: align API and dashboard coverage with routes
