+++ packages/junior-dashboard/e2e/people.spec.ts
@@ -1,0 +1,66 @@
+import { expect, test } from "@playwright/test";
+import {
+  collectBrowserErrors,
+  type DashboardE2eServer,
+  mockDashboardApis,
+  startDashboardE2eServer,
+} from "./harness";
+
+let server: DashboardE2eServer;
+
+test.beforeAll(async () => {
+  server = await startDashboardE2eServer();
+});
+
+test.afterAll(async () => {
+  await server.close();
+});
+
+test.beforeEach(async ({ page }) => {
+  await mockDashboardApis(page);
+});
+
+test("explores people activity", async ({ page }) => {
+  await page.setViewportSize({ height: 900, width: 1600 });
+  const browserErrors = collectBrowserErrors(page);
+  await page.goto(`${server.baseURL}/people`);
+
+  await expect(
+    page.locator("main > div").getByText("People", { exact: true }),
+  ).toBeVisible();
+  await expect(
+    page.getByRole("img", { name: "Active people per day" }),
+  ).toBeVisible();
+  await expect(page.getByRole("button", { name: "90d" })).toHaveAttribute(
+    "aria-pressed",
+    "true",
+  );
+  await expect(page.getByRole("combobox", { name: "Sort people" })).toHaveValue(
+    "conversations",
+  );
+  await page.getByRole("button", { name: "7d" }).click();
+  await expect(page.getByRole("button", { name: "7d" })).toHaveAttribute(
+    "aria-pressed",
+    "true",
+  );
+  await page.getByRole("button", { name: "90d" }).click();
+  await expect(page.getByRole("button", { name: "90d" })).toHaveAttribute(
+    "aria-pressed",
+    "true",
+  );
+
+  const headerBounds = await page
+    .locator("main > header > div")
+    .evaluate((element) => {
+      const bounds = element.getBoundingClientRect();
+      return { left: bounds.left, width: bounds.width };
+    });
+  const containerBounds = await page
+    .locator("main > div")
+    .evaluate((element) => {
+      const bounds = element.getBoundingClientRect();
+      return { left: bounds.left, width: bounds.width };
+    });
+  expect(containerBounds).toEqual(headerBounds);
+  expect(browserErrors).toEqual([]);
+});
