+++ packages/junior/src/api/runtime.ts
@@ -1,0 +1,11 @@
+import { readRuntimeInfoReport } from "../reporting";
+import { defineApiRoute } from "./route";
+import { runtimeInfoReportSchema } from "./schema";
+
+/** Serve Junior runtime build and environment metadata. */
+export default defineApiRoute({
+  method: "get",
+  path: "/api/runtime",
+  responseSchema: runtimeInfoReportSchema,
+  handler: readRuntimeInfoReport,
+});
