+++ packages/junior/tests/integration/api/health.test.ts
@@ -1,0 +1,14 @@
+import { describe, expect, test } from "vitest";
+import { createJuniorApi } from "@/api";
+import { healthReportSchema } from "@/api/schema";
+
+describe("health API route", () => {
+  test("serves the authenticated health report", async () => {
+    const response = await createJuniorApi().request(
+      "http://localhost/api/health",
+    );
+
+    expect(response.status).toBe(200);
+    healthReportSchema.parse(await response.json());
+  });
+});
